feat: refactor Unilogin client integration to fetch configuration dynamically
