+++ app/auth/callback/unilogin/route.ts
@@ -14,8 +14,20 @@ export interface TIntrospectionResponse extends IntrospectionResponse {
 }
 
 export async function GET(request: NextRequest) {
+  const configResponse = await fetch(`${process.env.NEXT_PUBLIC_APP_URL}/auth/config`)
+  if (!configResponse.ok) {
+    throw new Error("Failed to fetch config")
+  }
+  const config = await configResponse.json()
+  const uniloginConfig = config?.dplConfiguration?.unilogin
+
   const session = await getSession()
-  const client = await getUniloginClient()
+  const client = await getUniloginClient({
+    client_id: uniloginConfig.unilogin_api_client_id,
+    client_secret: uniloginConfig.unilogin_api_client_secret,
+    redirect_uri: `${process.env.NEXT_PUBLIC_APP_URL}/auth/callback/unilogin`,
+    wellKnownUrl: uniloginConfig.unilogin_api_wellknown_url,
+  })
   const params = client.callbackParams(request.nextUrl.toString())
 
   // Fetch all user/token info.
