+++ app/auth/login/unilogin/route.ts
@@ -1,46 +1,20 @@
-import { Issuer, generators } from "openid-client"
+import { generators } from "openid-client"
 
+import { getUniloginClient } from "@/lib/session/oauth/uniloginClient"
 import { getSession } from "@/lib/session/session"
 
 export const revalidate = 1
-
-export async function getUniloginClient({
-  client_id,
-  client_secret,
-  redirect_uri,
-  wellKnownUrl,
-}: {
-  client_id: string
-  client_secret: string
-  redirect_uri: string
-  wellKnownUrl: string
-}) {
-  const UniloginIssuer = await Issuer.discover(wellKnownUrl!)
-  const client = new UniloginIssuer.Client({
-    client_id: client_id!,
-    client_secret: client_secret!,
-    response_types: ["code"],
-    redirect_uris: [redirect_uri],
-    token_endpoint_auth_method: "client_secret_post",
-  })
-  return client
-}
 
 export async function GET() {
   const configResponse = await fetch(`${process.env.NEXT_PUBLIC_APP_URL}/auth/config`)
   if (!configResponse.ok) {
     throw new Error("Failed to fetch config")
   }
-
   const config = await configResponse.json()
-
   const uniloginConfig = config?.dplConfiguration?.unilogin
 
   const session = await getSession()
-
   session.code_verifier = generators.codeVerifier()
-
-  const code_challenge = generators.codeChallenge(session.code_verifier)
 
   const client = await getUniloginClient({
     client_id: uniloginConfig.unilogin_api_client_id,
@@ -48,6 +22,8 @@ export async function GET() {
     redirect_uri: `${process.env.NEXT_PUBLIC_APP_URL}/auth/callback/unilogin`,
     wellKnownUrl: uniloginConfig.unilogin_api_wellknown_url,
   })
+
+  const code_challenge = generators.codeChallenge(session.code_verifier)
   const url = client.authorizationUrl({
     scope: "openid",
     audience: process.env.UNILOGIN_API_URL,
